test: M6 real GGUF weights end-to-end test passes on gfx1151
- tests/test_mtp_nextn_real_gguf.py: loads real blk.40 NextN tensors from
  /models/gguf/Qwen3.6-35B-A3B-UD-Q4_K_M.gguf and runs the composite
  mtp_nextn_layer with real weights (Q8_0 eh_proj+attn, Q4_K gate/up, Q5_K down,
  Q8_0 shared, BF16 router, F32 norms, real shapes hidden=2048 heads=16 kv=2).
- Gates GPU output vs cpu_reference on the same real weights (max_abs < 1e-2).
  cpu_reference gets dequanted F32 for eh_proj+attn; GPU gets raw Q8_0 block
  bytes with qtype kwargs. MoE expert + shared weights use raw bytes for both.
- Fix: eh_proj weight was being force-cast to float32, corrupting Q8_0 uint8
  data. Changed to np.ascontiguousarray(eh_proj_weight) (preserve raw bytes).
- 48/48 tests pass on gfx1151. First end-to-end real-model GGUF MTP NextN
  forward on native GPU (hip_gfx1100/gfx1151).

+++ hipengine/kernels/hip_gfx1100/speculative/mtp_nextn.py
@@ -524,7 +524,7 @@ def qwen35_gguf_mtp_eh_proj_f32(
 
     hidden_arr = np.ascontiguousarray(hidden_seed, dtype=np.float32)
     embed_arr = np.ascontiguousarray(token_embedding, dtype=np.float32)
-    weight = np.ascontiguousarray(eh_proj_weight, dtype=np.float32)
+    weight = np.ascontiguousarray(eh_proj_weight)  # keep raw bytes for K-quant
     hnorm = np.ascontiguousarray(hnorm_weight, dtype=np.float32)
     enorm = np.ascontiguousarray(enorm_weight, dtype=np.float32)
     if hidden_arr.ndim != 2:
